requirements.txt
# Core
Django==5.2.8
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.11
redis==5.0.1

# API Documentation
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1

# Authentication
djangorestframework-simplejwt==5.5.1
python-jose==3.3.0
python-dotenv==1.2.1

# Security
django-cors-headers==4.9.0
django-ratelimit==4.1.0

# Monitoring
django-prometheus==2.3.1
prometheus-client==0.20.0

# Utilities
dj-database-url==3.0.1
python-dateutil==2.9.0
Pillow==11.0.0

# Development
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-filter
google-auth==2.37.0
requests==2.32.3
